refactor(server): move main router to `src/routes/`
